add Odyssey_Config_Data and changed unit tests
