+++ projects/packages/stats-admin/tests/php/test-stats-dashboard.php
@@ -8,7 +8,7 @@ use Automattic\Jetpack\Stats_Admin\Test_Case as Stats_Test_Case;
  *
  * @package automattic/jetpack-stats-admin
  */
-class Test_Plan extends Stats_Test_Case {
+class Test_Dashboard extends Stats_Test_Case {
 	/**
 	 * Test that init sets $initialized.
 	 */
@@ -37,36 +37,5 @@ class Test_Plan extends Stats_Test_Case {
 		$get_cdn_asset_cache_buster = new \ReflectionMethod( $dashboard, 'get_cdn_asset_cache_buster' );
 		$get_cdn_asset_cache_buster->setAccessible( true );
 		$this->assertEquals( 'calypso-4917-8664-g72a154d63a', $get_cdn_asset_cache_buster->invoke( $dashboard ) );
-	}
-
-	/**
-	 * Test configData set to JS.
-	 */
-	public function test_render_config_data() {
-		$dashboard   = new Dashboard();
-		$config_data = new \ReflectionMethod( $dashboard, 'get_config_data_js' );
-		$config_data->setAccessible( true );
-		$this->assertTrue( strpos( $config_data->invoke( $dashboard ), 'window.configData' ) === 0 );
-	}
-
-	/**
-	 * Test config_data has all necessary keys.
-	 */
-	public function test_config_data() {
-		$dashboard   = new Dashboard();
-		$config_data = new \ReflectionMethod( $dashboard, 'config_data' );
-		$config_data->setAccessible( true );
-		$data = $config_data->invoke( $dashboard );
-		$this->assertArrayHasKey( 'admin_page_base', $data );
-		$this->assertArrayHasKey( 'api_root', $data );
-		$this->assertArrayHasKey( 'blog_id', $data );
-		$this->assertArrayHasKey( 'env_id', $data );
-		$this->assertArrayHasKey( 'google_analytics_key', $data );
-		$this->assertArrayHasKey( 'google_maps_and_places_api_key', $data );
-		$this->assertArrayHasKey( 'i18n_default_locale_slug', $data );
-		$this->assertArrayHasKey( 'nonce', $data );
-		$this->assertArrayHasKey( 'site_name', $data );
-		$this->assertArrayHasKey( 'intial_state', $data );
-		$this->assertEmpty( $data['features'] );
 	}
 }
